补充完善接口 修复仓库 count 方法返回 Long 类型转换的问题

+++ server/src/main/java/org/tcpx/sharine/repository/CategoryRepository.java
@@ -19,5 +19,5 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     @Query("SELECT COUNT(c) FROM Category c WHERE c.title LIKE %:title%")
     @Cacheable(unless = "#result==null")
-    Long countByTitleRegex(String title);
+    Integer countByTitleRegex(String title);
 }
